Philipp Schöneberg Abgabe 1 (#3)
* Abgabe 1

* Improvement of the exercises

I changed the jupyter notebook file back and hope that is isn't a change anymore.  I also changed my printed strings to f-strings and used the recommended set functionality for the list comparison task which we spoke about in the lecture.

* Delete Vorlesungsunterlagen/02_Freitag.ipynb

* Improvements based on the given comments

line 11: typo 'givin'; capitalization 'fahrenheit'
line 15: count of the open and closed bracket as new variables
line 27, 28: renaming the variables in english
line 29, 31: changing i into elem to make the usage more clear

* Create 02_Freitag.ipynb

* Fixing the changed Jupyter Notebook

I deleted it from this branch, downloaded it again from the main branch and pushed it now again. So there shouldn't be a changed document besides my exercises in the commits.